extracted the redis monitor ot a different monitoring type
